requirements.txt
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
ecdsa==0.15
Flask==1.1.2
Flask-SocketIO==4.3.0
Flask-SQLAlchemy==2.4.2
future==0.18.2
idna==2.9
iso8601==0.1.12
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
pid==3.0.3
Pillow==7.1.2
pip==20.1.1
psutil==5.7.0
pyaes==1.6.1
pyserial==3.4
python-dotenv==0.13.0
python-engineio==3.13.0
python-socketio==4.6.0
PyYAML==5.3.1
requests==2.23.0
serial==0.0.97
setuptools==47.1.1
six==1.15.0
SQLAlchemy==1.3.17
-e git+https://github.com/texx00/sandypi.git@2d052c565c6e4a3e61bf169ed776b68f72ed49b9#egg=UIserver
urllib3==1.25.9
Werkzeug==1.0.1
